# shared/contracts/openapi/rpml-web-app.v1.yaml
openapi: 3.1.0
info:
  title: RPML Backend
  description: 'RPML Web App HTTP API (contract snapshot: shared/contracts/openapi/rpml-web-app.v1.yaml).'
  version: 1.0.0
paths:
...
  /api/v1/optimization/tasks:
    post:
      tags:
      - optimization
      summary: Создать асинхронную задачу оптимизации
      description: Ставит расчет в очередь Celery и сразу возвращает task_id.
      operationId: create_optimization_task_api_v1_optimization_tasks_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOptimizationTaskRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOptimizationTaskResponse'
        '401':
          description: Пользователь не аутентифицирован.
        '422':
          description: Unprocessable Content
          $ref: '#/components/responses/ErrorContent'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    CreateOptimizationTaskRequest:
      properties:
        input_mode:
          type: string
          enum:
          - scenario_snapshot
          - dataset_instance
          title: Input Mode
          description: Источник входных данных для фоновой задачи.
          default: scenario_snapshot
        instance_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Name
          description: Имя `.dat` инстанса. Обязательно для `dataset_instance`.
        horizon_months:
          anyOf:
          - type: integer
            maximum: 360.0
            minimum: 1.0
          - type: 'null'
          title: Horizon Months
          description: 'Горизонт планирования. Для dataset mode можно не передавать:
            будет использован горизонт инстанса.'
        ru_mode:
          type: boolean
          title: Ru Mode
          description: Если true, применяются правила RU-режима.
          default: true
        mc_income:
          type: boolean
          title: Mc Income
          description: Если true, в фоне будет рассчитан Monte Carlo summary.
          default: false
        mc_config:
          anyOf:
          - $ref: '#/components/schemas/OptimizationMonteCarloConfig'
          - type: 'null'
          description: Параметры Monte Carlo для фонового расчета.
      type: object
      title: CreateOptimizationTaskRequest
    CreateOptimizationTaskResponse:
      properties:
        task_id:
          type: string
          title: Task Id
          description: Идентификатор асинхронной задачи.
        status:
          type: string
          const: pending
          title: Status
          default: pending
        input_mode:
          type: string
          enum:
          - scenario_snapshot
          - dataset_instance
          title: Input Mode
          description: Режим входных данных задачи.
        horizon_months:
          type: integer
          title: Horizon Months
          description: Фактический горизонт задачи.
        instance_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Name
          description: Имя dataset-инстанса, если задача создана в `dataset_instance`
            режиме.
        ru_mode:
          type: boolean
          title: Ru Mode
          description: RU-режим, зафиксированный в задаче.
        mc_income:
          type: boolean
          title: Mc Income
          description: MC-режим, зафиксированный в задаче.
      type: object
      required:
      - task_id
      - input_mode
      - horizon_months
      - ru_mode
      - mc_income
      title: CreateOptimizationTaskResponse
...
    OptimizationMonteCarloConfig:
      properties:
        n_scenarios:
          type: integer
          minimum: 1.0
          title: N Scenarios
          description: Количество сценариев Monte Carlo.
          default: 16
        seed:
          type: integer
          title: Seed
          description: Базовый seed генератора случайных траекторий.
          default: 42
        rho:
          type: number
          maximum: 0.999
          minimum: -0.999
          title: Rho
          description: AR(1) коэффициент автокорреляции дохода.
          default: 0.55
        sigma:
          type: number
          minimum: 0.0
          title: Sigma
          description: Волатильность логнормального шума.
          default: 0.15
        shock_prob:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Shock Prob
          description: Вероятность шокового снижения дохода в месяце.
          default: 0.04
        shock_severity_mean:
          type: number
          minimum: 0.0
          title: Shock Severity Mean
          description: Средняя глубина шока дохода.
          default: 0.3
        shock_severity_std:
          type: number
          minimum: 0.0
          title: Shock Severity Std
          description: Стандартное отклонение глубины шока дохода.
          default: 0.1
        min_income_floor:
          type: number
          minimum: 0.0
          title: Min Income Floor
          description: Нижняя отсечка месячного дохода.
          default: 1.0
      type: object
      title: OptimizationMonteCarloConfig
...
    ErrorResponse:
      type: object
      required:
      - detail
      title: ErrorResponse
      properties:
        detail:
          description: 'Error detail: string from HTTPException or list for validation
            errors.'
          oneOf:
          - type: string
          - type: array
            items: {}
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
  responses:
    ErrorContent:
      description: Error body for many 4xx/5xx responses (HTTPException).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'